+++ src/lib.rs
@@ -5,6 +5,7 @@ pub mod obs;
 pub mod obs_routing;
 pub mod raffle;
 pub mod sdf_effects;
+pub mod stream_character;
 pub mod stream_fx;
 pub mod themesong;
 pub mod uberduck;
